Extract shared free-space parser utility
